pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "agent-world-compiler-poc"
version = "0.1.0"
description = "Minimal PoC: observe agent execution → derive capability profile → compile World Manifest → enforce deterministic decisions."
readme = "README.md"
requires-python = ">=3.12"
license = { text = "MIT" }
authors = [{ name = "Sergey Vlasov" }]
dependencies = [
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
]

[project.scripts]
awc-evaluate = "awc.policy.evaluate:main"

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v"

[tool.setuptools.packages.find]
where = ["src"]
include = ["awc*"]
